LayoutTests/imported/w3c/web-platform-tests/wasm/webapi/esm-integration/worker.tentative-expected.txt

CONSOLE MESSAGE: Error: assert_unreached: worker got an error Reached unreachable code
CONSOLE MESSAGE: SyntaxError: Invalid character: '\0'

FAIL Testing WebAssembly worker Error: assert_unreached: worker got an error Reached unreachable code
